+++ lib/admin/route/admin.flow/admin.flow.controller.js
@@ -6,7 +6,7 @@ const {
   deepClone
 } = require('@ministryofjustice/fb-utils-node')
 
-const {propagateInstanceConditions} = require('~/fb-editor-node/lib/runtime/propagate-show')
+const {propagateInstanceConditions} = require('~/fb-editor-node/lib/service-data/runtime/propagate-show')
 
 const {
   getEntryPointInstances,
